refactor: 상품 관리 페이지에 이벤트를 추가하는 부분을 ProductAddEvent.js로 분리
