+++ docs/README.md
@@ -12,12 +12,12 @@
   - [x] 3개의 탭을 화면에 표시한다
   - [x] 탭을 누르면 해당하는 내용을 보여준다
   - [ ] 기존 탭의 상태가 유지되어야 한다
-  - [ ] 새로고침해도 가장 최근의 정보들을 불러와야 한다
-  - [ ] localStorage에 처음 값 들어갔을 때 null로 표시되는 오류 수정
+  - [x] 새로고침해도 가장 최근의 정보들을 불러와야 한다
+  - [x] localStorage에 처음 값 들어갔을 때 null로 표시되는 오류 수정
 
 <br>
 
-- [ ] 상품 관리 탭
+- [x] 상품 관리 탭
   - [x] 상품을 추가할 수 있어야 한다
   - [x] 추가한 상품을 화면에 표시한다
   - [x] 상품 가격이 100원 이상인지 검증한다
@@ -25,7 +25,7 @@
 
 <br>
 
-- [ ] 잔돈 충전 탭
+- [x] 잔돈 충전 탭
   - [x] 충전하기 버튼을 눌러 자판기 보유 금액을 누적하여 충전한다
   - [x] 자판기 보유 금액 만큼의 동전을 랜덤으로 생성한다
   - [x] 자판기 보유 총 금액을 화면에 표시한다
@@ -33,7 +33,7 @@
 
 <br>
 
-- [ ] 상품 구매 탭
+- [x] 상품 구매 탭
   - [x] 금액을 입력받아 누적하여 충전한다
   - [x] 금액이 10원으로 나누어 떨어지는지 검증한다
   - [x] 총 투입된 금액을 화면에 표시한다
